requirements.txt
torch
transformers
accelerate
pyboy
pillow
sentencepiece
qwen-vl-utils
